refactor(storage): move `PerformUpload`
Move `PerformUpload` to a separate module, with separate tests modules.
