+++ src/storage/src/storage.rs
@@ -15,6 +15,7 @@
 #[allow(dead_code)]
 pub(crate) mod checksum;
 pub(crate) mod client;
+pub(crate) mod perform_upload;
 pub(crate) mod read_object;
 pub(crate) mod request_options;
 pub(crate) mod upload_object;
@@ -22,7 +23,7 @@ pub mod upload_source;
 pub(crate) mod v1;
 
 use crate::model::Object;
-use crate::upload_source::{InsertPayload, StreamingSource};
+use crate::upload_source::InsertPayload;
 use crate::{Error, Result};
 
 /// An unrecoverable problem in the upload protocol.
