+++ src/storage/src/storage/perform_upload/buffered.rs
@@ -12,41 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod progress;
-
-use super::*;
-use crate::{retry_policy::ContinueOn308, upload_source::IterSource};
+use super::{
+    ContinueOn308, Error, IterSource, Object, PerformUpload, Result, ResumableUploadStatus,
+    StreamingSource, X_GOOG_API_CLIENT_HEADER, apply_customer_supplied_encryption_headers,
+};
 use progress::InProgressUpload;
+use std::sync::Arc;
+use tokio::sync::Mutex;
 
-impl<T> UploadObject<T>
-where
-    T: StreamingSource + Send + Sync + 'static,
-    T::Error: std::error::Error + Send + Sync + 'static,
-{
-    /// Upload an object from a streaming source without rewinds.
-    ///
-    /// # Example
-    /// ```
-    /// # use google_cloud_storage::client::Storage;
-    /// # async fn sample(client: &Storage) -> anyhow::Result<()> {
-    /// let response = client
-    ///     .upload_object("projects/_/buckets/my-bucket", "my-object", "hello world")
-    ///     .send()
-    ///     .await?;
-    /// println!("response details={response:?}");
-    /// # Ok(()) }
-    /// ```
-    pub async fn send(self) -> crate::Result<Object> {
-        self.build().send().await
-    }
-}
+mod progress;
 
 impl<S> PerformUpload<S>
 where
     S: StreamingSource + Send + Sync + 'static,
     S::Error: std::error::Error + Send + Sync + 'static,
 {
-    async fn send(self) -> crate::Result<Object> {
+    pub(crate) async fn send(self) -> crate::Result<Object> {
         let hint = self
             .payload
             .lock()
@@ -138,7 +119,7 @@ where
             .header("Content-Range", range)
             .header(
                 "x-goog-api-client",
-                reqwest::header::HeaderValue::from_static(&self::info::X_GOOG_API_CLIENT_HEADER),
+                reqwest::header::HeaderValue::from_static(&X_GOOG_API_CLIENT_HEADER),
             );
 
         let builder = apply_customer_supplied_encryption_headers(builder, &self.params);
@@ -172,8 +153,12 @@ mod resumable_tests;
 
 #[cfg(test)]
 mod tests {
-    use super::super::client::tests::{test_builder, test_inner_client};
     use super::*;
+    use crate::storage::client::{
+        Storage,
+        tests::{test_builder, test_inner_client},
+    };
+    use crate::storage::upload_object::UploadObject;
     use httptest::{Expectation, Server, matchers::*, responders::status_code};
     use test_case::test_case;
 
@@ -287,20 +272,6 @@ mod tests {
             })
             .unwrap();
         assert_eq!(got, want);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn handle_start_resumable_upload_response() -> Result {
-        let response = http::Response::builder()
-            .header(
-                "Location",
-                "http://private.googleapis.com/test-only/session-123",
-            )
-            .body(Vec::new())?;
-        let response = reqwest::Response::from(response);
-        let url = super::handle_start_resumable_upload_response(response).await?;
-        assert_eq!(url, "http://private.googleapis.com/test-only/session-123");
         Ok(())
     }
 }
