+++ src/storage/src/storage/upload_object.rs
@@ -13,12 +13,9 @@
 // limitations under the License.
 
 use super::client::*;
+use super::perform_upload::PerformUpload;
+use super::upload_source::{Seek, StreamingSource};
 use super::*;
-use std::sync::Arc;
-use tokio::sync::Mutex;
-
-mod buffered;
-mod unbuffered;
 
 /// A request builder for uploads without rewind.
 pub struct UploadObject<T> {
@@ -26,24 +23,6 @@ pub struct UploadObject<T> {
     spec: crate::model::WriteObjectSpec,
     params: Option<crate::model::CommonObjectRequestParams>,
     payload: InsertPayload<T>,
-    options: super::request_options::RequestOptions,
-}
-
-/// Represents an upload constructed via `UploadObject<T>`.
-///
-/// Once the application has fully configured an `UploadObject<T>` it calls
-/// `send()` or `send_buffered()` to initiate the upload. At that point the
-/// client library creates an instance of this class. Notable, the `payload`
-/// becomes `Arc<Mutex<T>>` because it needs to be reused in the retry loop.
-///
-/// TODO(#2050) - payload will become `ChecksummedSource<C, S>` to automatically
-///   compute the checksum.
-struct PerformUpload<S> {
-    // We need `Arc<Mutex<>>` because this is re-used in retryable uploads.
-    payload: Arc<Mutex<S>>,
-    inner: std::sync::Arc<StorageInner>,
-    spec: crate::model::WriteObjectSpec,
-    params: Option<crate::model::CommonObjectRequestParams>,
     options: super::request_options::RequestOptions,
 }
 
@@ -548,7 +527,7 @@ impl<T> UploadObject<T> {
 
     // TODO(#2050) - this should be automatically computed?
     #[allow(dead_code)]
-    fn with_crc32c<V>(mut self, v: V) -> Self
+    pub(crate) fn with_crc32c<V>(mut self, v: V) -> Self
     where
         V: Into<u32>,
     {
@@ -560,7 +539,7 @@ impl<T> UploadObject<T> {
 
     // TODO(#2050) - this should be automatically computed?
     #[allow(dead_code)]
-    fn with_md5_hash<I, V>(mut self, i: I) -> Self
+    pub(crate) fn with_md5_hash<I, V>(mut self, i: I) -> Self
     where
         I: IntoIterator<Item = V>,
         V: Into<u8>,
@@ -754,197 +733,69 @@ impl<T> UploadObject<T> {
         }
     }
 
-    fn build(self) -> PerformUpload<InsertPayload<T>> {
-        PerformUpload {
-            payload: Arc::new(Mutex::new(self.payload)),
-            inner: self.inner,
-            spec: self.spec,
-            params: self.params,
-            options: self.options,
-        }
-    }
-}
-
-impl<S> PerformUpload<S> {
-    fn resource(&self) -> &crate::model::Object {
-        self.spec
-            .resource
-            .as_ref()
-            .expect("resource field initialized in `new()`")
-    }
-
-    async fn start_resumable_upload_attempt(&self) -> Result<String> {
-        let builder = self.start_resumable_upload_request().await?;
-        let response = builder.send().await.map_err(Error::io)?;
-        self::handle_start_resumable_upload_response(response).await
-    }
-
-    async fn start_resumable_upload_request(&self) -> Result<reqwest::RequestBuilder> {
-        let bucket = &self.resource().bucket;
-        let bucket_id = bucket.strip_prefix("projects/_/buckets/").ok_or_else(|| {
-            Error::binding(format!(
-                "malformed bucket name, it must start with `projects/_/buckets/`: {bucket}"
-            ))
-        })?;
-        let object = &self.resource().name;
-        let builder = self
-            .inner
-            .client
-            .request(
-                reqwest::Method::POST,
-                format!("{}/upload/storage/v1/b/{bucket_id}/o", &self.inner.endpoint),
-            )
-            .query(&[("uploadType", "resumable")])
-            .query(&[("name", enc(object))])
-            .header("content-type", "application/json")
-            .header(
-                "x-goog-api-client",
-                reqwest::header::HeaderValue::from_static(&self::info::X_GOOG_API_CLIENT_HEADER),
-            );
-
-        let builder = self.apply_preconditions(builder);
-        let builder = apply_customer_supplied_encryption_headers(builder, &self.params);
-        let builder = self.inner.apply_auth_headers(builder).await?;
-        let builder = builder.json(&v1::insert_body(self.resource()));
-        Ok(builder)
-    }
-
-    async fn query_resumable_upload_attempt(
-        &self,
-        upload_url: &str,
-    ) -> Result<ResumableUploadStatus> {
-        let builder = self
-            .inner
-            .client
-            .request(reqwest::Method::PUT, upload_url)
-            .header("content-type", "application/octet-stream")
-            .header("Content-Range", "bytes */*")
-            .header(
-                "x-goog-api-client",
-                reqwest::header::HeaderValue::from_static(&self::info::X_GOOG_API_CLIENT_HEADER),
-            );
-        let builder = self.inner.apply_auth_headers(builder).await?;
-        let response = builder.send().await.map_err(Error::io)?;
-        self::query_resumable_upload_handle_response(response).await
-    }
-
-    fn apply_preconditions(&self, builder: reqwest::RequestBuilder) -> reqwest::RequestBuilder {
-        let builder = self
-            .spec
-            .if_generation_match
-            .iter()
-            .fold(builder, |b, v| b.query(&[("ifGenerationMatch", v)]));
-        let builder = self
-            .spec
-            .if_generation_not_match
-            .iter()
-            .fold(builder, |b, v| b.query(&[("ifGenerationNotMatch", v)]));
-        let builder = self
-            .spec
-            .if_metageneration_match
-            .iter()
-            .fold(builder, |b, v| b.query(&[("ifMetagenerationMatch", v)]));
-        let builder = self
-            .spec
-            .if_metageneration_not_match
-            .iter()
-            .fold(builder, |b, v| b.query(&[("ifMetagenerationNotMatch", v)]));
-
-        [
-            ("kmsKeyName", self.resource().kms_key.as_str()),
-            ("predefinedAcl", self.spec.predefined_acl.as_str()),
-        ]
-        .into_iter()
-        .fold(
-            builder,
-            |b, (k, v)| if v.is_empty() { b } else { b.query(&[(k, v)]) },
+    pub(crate) fn build(self) -> PerformUpload<InsertPayload<T>> {
+        PerformUpload::new(
+            self.payload,
+            self.inner,
+            self.spec,
+            self.params,
+            self.options,
         )
     }
 }
 
-async fn handle_start_resumable_upload_response(response: reqwest::Response) -> Result<String> {
-    if !response.status().is_success() {
-        return gaxi::http::to_http_error(response).await;
-    }
-    let location = response
-        .headers()
-        .get("Location")
-        .ok_or_else(|| Error::deser("missing Location header in start resumable upload"))?;
-    location.to_str().map_err(Error::deser).map(str::to_string)
-}
-
-async fn query_resumable_upload_handle_response(
-    response: reqwest::Response,
-) -> Result<ResumableUploadStatus> {
-    if response.status() == RESUME_INCOMPLETE {
-        return self::parse_range(response).await;
-    }
-    let object = handle_object_response(response).await?;
-    Ok(ResumableUploadStatus::Finalized(Box::new(object)))
-}
-
-async fn handle_object_response(response: reqwest::Response) -> Result<Object> {
-    if !response.status().is_success() {
-        return gaxi::http::to_http_error(response).await;
-    }
-    let response = response.json::<v1::Object>().await.map_err(Error::deser)?;
-    Ok(Object::from(response))
-}
-
-async fn parse_range(response: reqwest::Response) -> Result<ResumableUploadStatus> {
-    let Some(end) = self::parse_range_end(response.headers()) else {
-        return gaxi::http::to_http_error(response).await;
-    };
-    // The `Range` header returns an inclusive range, i.e. bytes=0-999 means "1000 bytes".
-    let persisted_size = match end {
-        0 => 0,
-        e => e + 1,
-    };
-    Ok(ResumableUploadStatus::Partial(persisted_size))
-}
-
-#[derive(Debug, PartialEq)]
-enum ResumableUploadStatus {
-    Finalized(Box<Object>),
-    Partial(u64),
-}
-
-fn parse_range_end(headers: &reqwest::header::HeaderMap) -> Option<u64> {
-    let Some(range) = headers.get("range") else {
-        // A missing `Range:` header indicates that no bytes are persisted.
-        return Some(0_u64);
-    };
-    // Uploads must be sequential, so the persisted range (if present) always
-    // starts at zero. This is poorly documented, but can be inferred from
-    //   https://cloud.google.com/storage/docs/performing-resumable-uploads#resume-upload
-    // which requires uploads to continue from the last byte persisted. It is
-    // better documented in the gRPC version, where holes are explicitly
-    // forbidden:
-    //   https://github.com/googleapis/googleapis/blob/302273adb3293bb504ecd83be8e1467511d5c779/google/storage/v2/storage.proto#L1253-L1255
-    let end = std::str::from_utf8(range.as_bytes().strip_prefix(b"bytes=0-")?).ok()?;
-    end.parse::<u64>().ok()
-}
-
-fn send_err(error: reqwest::Error) -> Error {
-    match error {
-        e if e.is_body() => Error::ser(e),
-        e if e.is_request() => Error::ser(e),
-        e if e.is_timeout() => Error::timeout(e),
-        e => Error::io(e),
+impl<T> UploadObject<T>
+where
+    T: StreamingSource + Seek + Send + Sync + 'static,
+    <T as StreamingSource>::Error: std::error::Error + Send + Sync + 'static,
+    <T as Seek>::Error: std::error::Error + Send + Sync + 'static,
+{
+    /// A simple upload from a buffer.
+    ///
+    /// # Example
+    /// ```
+    /// # use google_cloud_storage::client::Storage;
+    /// # async fn sample(client: &Storage) -> anyhow::Result<()> {
+    /// let response = client
+    ///     .upload_object("projects/_/buckets/my-bucket", "my-object", "hello world")
+    ///     .send_unbuffered()
+    ///     .await?;
+    /// println!("response details={response:?}");
+    /// # Ok(()) }
+    /// ```
+    pub async fn send_unbuffered(self) -> Result<Object> {
+        self.build().send_unbuffered().await
     }
 }
 
-const RESUME_INCOMPLETE: reqwest::StatusCode = reqwest::StatusCode::PERMANENT_REDIRECT;
+impl<T> UploadObject<T>
+where
+    T: StreamingSource + Send + Sync + 'static,
+    T::Error: std::error::Error + Send + Sync + 'static,
+{
+    /// Upload an object from a streaming source without rewinds.
+    ///
+    /// # Example
+    /// ```
+    /// # use google_cloud_storage::client::Storage;
+    /// # async fn sample(client: &Storage) -> anyhow::Result<()> {
+    /// let response = client
+    ///     .upload_object("projects/_/buckets/my-bucket", "my-object", "hello world")
+    ///     .send()
+    ///     .await?;
+    /// println!("response details={response:?}");
+    /// # Ok(()) }
+    /// ```
+    pub async fn send(self) -> crate::Result<Object> {
+        self.build().send().await
+    }
+}
 
 #[cfg(test)]
 mod tests {
-    use super::client::tests::{create_key_helper, test_builder, test_inner_client};
+    use super::client::tests::{test_builder, test_inner_client};
     use super::*;
     use crate::model::WriteObjectSpec;
-
-    use serde_json::{Value, json};
-    use std::collections::BTreeMap;
-    use test_case::test_case;
 
     type Result = anyhow::Result<()>;
 
@@ -1118,189 +969,5 @@ mod tests {
             .with_resumable_upload_buffer_size(456_usize);
         assert_eq!(request.options.resumable_upload_threshold, 345);
         assert_eq!(request.options.resumable_upload_buffer_size, 456);
-    }
-
-    #[tokio::test]
-    async fn start_resumable_upload() -> Result {
-        let inner = test_inner_client(test_builder());
-        let mut request = UploadObject::new(inner, "projects/_/buckets/bucket", "object", "hello")
-            .build()
-            .start_resumable_upload_request()
-            .await?
-            .build()?;
-
-        assert_eq!(request.method(), reqwest::Method::POST);
-        assert_eq!(
-            request.url().as_str(),
-            "http://private.googleapis.com/upload/storage/v1/b/bucket/o?uploadType=resumable&name=object"
-        );
-        let body = request.body_mut().take().unwrap();
-        let contents = http_body_util::BodyExt::collect(body).await?.to_bytes();
-        let json = serde_json::from_slice::<Value>(&contents)?;
-        assert_eq!(json, json!({}));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn start_resumable_upload_headers() -> Result {
-        // Make a 32-byte key.
-        let (key, key_base64, _, key_sha256_base64) = create_key_helper();
-
-        let inner = test_inner_client(test_builder());
-        let request = UploadObject::new(inner, "projects/_/buckets/bucket", "object", "hello")
-            .with_key(KeyAes256::new(&key)?)
-            .build()
-            .start_resumable_upload_request()
-            .await?
-            .build()?;
-
-        assert_eq!(request.method(), reqwest::Method::POST);
-        assert_eq!(
-            request.url().as_str(),
-            "http://private.googleapis.com/upload/storage/v1/b/bucket/o?uploadType=resumable&name=object"
-        );
-
-        let want = vec![
-            ("x-goog-encryption-algorithm", "AES256".to_string()),
-            ("x-goog-encryption-key", key_base64),
-            ("x-goog-encryption-key-sha256", key_sha256_base64),
-        ];
-
-        for (name, value) in want {
-            assert_eq!(
-                request.headers().get(name).unwrap().as_bytes(),
-                bytes::Bytes::from(value)
-            );
-        }
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn start_resumable_upload_bad_bucket() -> Result {
-        let inner = test_inner_client(test_builder());
-        UploadObject::new(inner, "malformed", "object", "hello")
-            .build()
-            .start_resumable_upload_request()
-            .await
-            .expect_err("malformed bucket string should error");
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn start_resumable_upload_metadata_in_request() -> Result {
-        use crate::model::ObjectAccessControl;
-        let inner = test_inner_client(test_builder());
-        let mut request = UploadObject::new(inner, "projects/_/buckets/bucket", "object", "")
-            .with_if_generation_match(10)
-            .with_if_generation_not_match(20)
-            .with_if_metageneration_match(30)
-            .with_if_metageneration_not_match(40)
-            .with_predefined_acl("private")
-            .with_acl([ObjectAccessControl::new()
-                .set_entity("allAuthenticatedUsers")
-                .set_role("READER")])
-            .with_cache_control("public; max-age=7200")
-            .with_content_disposition("inline")
-            .with_content_encoding("gzip")
-            .with_content_language("en")
-            .with_content_type("text/plain")
-            .with_crc32c(crc32c::crc32c(b""))
-            .with_custom_time(wkt::Timestamp::try_from("2025-07-07T18:11:00Z")?)
-            .with_event_based_hold(true)
-            .with_md5_hash(md5::compute(b"").0)
-            .with_metadata([("k0", "v0"), ("k1", "v1")])
-            .with_retention(
-                crate::model::object::Retention::new()
-                    .set_mode(crate::model::object::retention::Mode::Locked)
-                    .set_retain_until_time(wkt::Timestamp::try_from("2035-07-07T18:14:00Z")?),
-            )
-            .with_storage_class("ARCHIVE")
-            .with_temporary_hold(true)
-            .with_kms_key("test-key")
-            .build()
-            .start_resumable_upload_request()
-            .await?
-            .build()?;
-
-        assert_eq!(request.method(), reqwest::Method::POST);
-        let want_pairs: BTreeMap<String, String> = [
-            ("uploadType", "resumable"),
-            ("name", "object"),
-            ("ifGenerationMatch", "10"),
-            ("ifGenerationNotMatch", "20"),
-            ("ifMetagenerationMatch", "30"),
-            ("ifMetagenerationNotMatch", "40"),
-            ("kmsKeyName", "test-key"),
-            ("predefinedAcl", "private"),
-        ]
-        .iter()
-        .map(|(k, v)| (k.to_string(), v.to_string()))
-        .collect();
-        let query_pairs: BTreeMap<String, String> = request
-            .url()
-            .query_pairs()
-            .map(|param| (param.0.to_string(), param.1.to_string()))
-            .collect();
-        assert_eq!(query_pairs, want_pairs);
-
-        let body = request.body_mut().take().unwrap();
-        let contents = http_body_util::BodyExt::collect(body).await?.to_bytes();
-        let json = serde_json::from_slice::<Value>(&contents)?;
-        assert_eq!(
-            json,
-            json!({
-                "acl": [{"entity": "allAuthenticatedUsers", "role": "READER"}],
-                "cacheControl": "public; max-age=7200",
-                "contentDisposition": "inline",
-                "contentEncoding": "gzip",
-                "contentLanguage": "en",
-                "contentType": "text/plain",
-                "crc32c": "AAAAAA==",
-                "customTime": "2025-07-07T18:11:00Z",
-                "eventBasedHold": true,
-                "md5Hash": "1B2M2Y8AsgTpgAmY7PhCfg==",
-                "metadata": {"k0": "v0", "k1": "v1"},
-                "retention": {"mode": "LOCKED", "retainUntilTime": "2035-07-07T18:14:00Z"},
-                "storageClass": "ARCHIVE",
-                "temporaryHold": true,
-            })
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn start_resumable_upload_credentials() -> Result {
-        let inner = test_inner_client(
-            test_builder().with_credentials(auth::credentials::testing::error_credentials(false)),
-        );
-        let _ = UploadObject::new(inner, "projects/_/buckets/bucket", "object", "hello")
-            .build()
-            .start_resumable_upload_request()
-            .await
-            .inspect_err(|e| assert!(e.is_authentication()))
-            .expect_err("invalid credentials should err");
-        Ok(())
-    }
-
-    #[test_case(None, Some(0))]
-    #[test_case(Some("bytes=0-12345"), Some(12345))]
-    #[test_case(Some("bytes=0-1"), Some(1))]
-    #[test_case(Some("bytes=0-0"), Some(0))]
-    #[test_case(Some("bytes=1-12345"), None)]
-    #[test_case(Some(""), None)]
-    fn range_end(input: Option<&str>, want: Option<u64>) {
-        use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
-        let headers = HeaderMap::from_iter(input.into_iter().map(|s| {
-            (
-                HeaderName::from_static("range"),
-                HeaderValue::from_str(s).unwrap(),
-            )
-        }));
-        assert_eq!(super::parse_range_end(&headers), want, "{headers:?}");
-    }
-
-    #[test]
-    fn validate_status_code() {
-        assert_eq!(RESUME_INCOMPLETE, 308);
     }
 }
